AdaptiveController_2015/4/19
The 3rd submission

Rename the class CCD to RealCCD.
Add a class FakeCCD to replace the class RealCCD without a real CCD.
Add functions to maintain a image desired in the class SLM, a image
desired in the class RealCCD and a image desired in the class FakeCCD.
Add functions to save the result of SA.
Add functions to time the SA process.
Add a function to interrupt the SA process.