AdaptiveController_2015/5/17
The 5th submission

Add functions in the class SLM, the class FakeCCD, the class RealCCD and
the class SimulatedAnnealing to process the events of the program.
Add functions in the class SLM, the class FakeCCD, the class RealCCD and
the class SimulatedAnnealing to save the log of each class.
Add functions in the class SimulatedAnnealing to save the data of the SA
process and the SA result.