AdaptiveController_2015/6/1
The 6th submission

Add functions in the class SLM to create, update and release a copy of
the phase out of the class.
Add functions in the class FakeCCD and the class RealCCD to create,
update and release a copy of the snapshot out of each class.
Add functions in the class SimulatedAnnealing to maintain the
imageCurrent as the processed CCD image.